requirements.txt
flask
mysql-connector-python
pymysql
pandas
